+++ src/app/lhc-form/lhc-form.component.ts
@@ -183,7 +183,7 @@ export class LhcFormComponent implements OnInit, OnChanges, OnDestroy, AfterView
                     })
                     .catch(error => {
                       const e = typeof error === "string" ? error : error.message
-                      self.onError.emit(e)
+                      self.onError.emit([e])
                     });
                 }
                 else {
@@ -191,8 +191,9 @@ export class LhcFormComponent implements OnInit, OnChanges, OnDestroy, AfterView
                 }
               })
               .catch(error => {
-                const e = typeof error === "string" ? error : error.message
-                self.onError.emit(e)
+                // loadFHIRResources can return an array of errors
+                const e = error.message ? error.message : error;
+                self.onError.emit([e])
               });
             }
             else {
@@ -202,7 +203,7 @@ export class LhcFormComponent implements OnInit, OnChanges, OnDestroy, AfterView
           }
           catch (error) {
             const e = typeof error === "string" ? error : error.message
-            self.onError.emit(e)
+            self.onError.emit([e])
           };
         },1)
       }